+++ packages/dotcom/src/index.ts
@@ -1,2 +1,3 @@
 export { getEpicViewLog } from '../../shared/src/lib/viewLog';
 export { getWeeklyArticleHistory } from '../../shared/src/lib/history';
+export * from './requests';
