--- packages/shared/src/types/epic/index.ts
@@ -1,3 +1,0 @@
-export * from './payload';
-export * from './test';
-export * from './props';
