+++ packages/shared/src/types/payloads/index.ts
@@ -1,0 +1,4 @@
+export * from './epic';
+export * from './header';
+export * from './banner';
+export * from './shared';
