+++ packages/shared/src/types/props/index.ts
@@ -1,0 +1,4 @@
+export * from './epic';
+export * from './banner';
+export * from './header';
+export * from './shared';
