--- packages/shared/src/types/shared/index.ts
@@ -1,3 +1,0 @@
-export * from './props';
-export * from './payload';
-export * from './test';
