+++ packages/shared/src/types/tests/banner.ts
@@ -1,45 +1,15 @@
-import * as z from 'zod';
-import { OphanProduct, OphanComponentType, OphanComponentEvent } from './ophan';
-import { CountryGroupId } from '../lib/geolocation';
+import { BannerChannel, Cta, SecondaryCta, TickerSettings } from '../props';
 import {
     ArticlesViewedSettings,
-    UserCohort,
-    Test,
-    TickerSettings,
-    Variant,
-    WeeklyArticleHistory,
     ControlProportionSettings,
-    ctaSchema,
-    secondaryCtaSchema,
-    tickerSettingsSchema,
-    Cta,
-    SecondaryCta,
-    Tracking,
-    PageTracking,
-    trackingSchema,
     TargetingAbTest,
+    Test,
+    UserCohort,
+    Variant,
 } from './shared';
-
-export type BannerTargeting = {
-    alreadyVisitedCount: number;
-    shouldHideReaderRevenue?: boolean;
-    isPaidContent?: boolean;
-    showSupportMessaging: boolean;
-    engagementBannerLastClosedAt?: string;
-    subscriptionBannerLastClosedAt?: string;
-    mvtId: number;
-    countryCode: string;
-    weeklyArticleHistory?: WeeklyArticleHistory;
-    hasOptedOutOfArticleCount: boolean;
-    modulesVersion?: string;
-    sectionId?: string;
-    tagIds?: string[];
-};
-
-export type BannerDataRequestPayload = {
-    tracking: PageTracking;
-    targeting: BannerTargeting;
-};
+import { OphanComponentType, OphanProduct } from '../ophan';
+import { BannerTargeting, PageTracking } from '../payloads';
+import { CountryGroupId } from '../../lib';
 
 export interface BannerContent {
     heading?: string;
@@ -49,15 +19,6 @@ export interface BannerContent {
     cta?: Cta;
     secondaryCta?: SecondaryCta;
 }
-
-const bannerContentSchema = z.object({
-    heading: z.string().optional(),
-    messageText: z.string(),
-    mobileMessageText: z.string().optional(),
-    highlightedText: z.string().optional(),
-    cta: ctaSchema.optional(),
-    secondaryCta: secondaryCtaSchema.optional(),
-});
 
 export enum BannerTemplate {
     ContributionsBanner = 'ContributionsBanner',
@@ -80,10 +41,6 @@ export interface BannerVariant extends Variant {
     componentType: OphanComponentType;
     products?: OphanProduct[];
 }
-
-export const bannerChannelSchema = z.enum(['contributions', 'subscriptions']);
-
-export type BannerChannel = z.infer<typeof bannerChannelSchema>;
 
 export type CanRun = (targeting: BannerTargeting, pageTracking: PageTracking) => boolean;
 
@@ -110,40 +67,6 @@ export interface BannerTestSelection {
     moduleUrl: string;
     moduleName: string;
     targetingAbTest?: TargetingAbTest;
-}
-
-export interface BannerProps {
-    tracking: Tracking;
-    bannerChannel: BannerChannel;
-    content?: BannerContent;
-    mobileContent?: BannerContent;
-    countryCode?: string;
-    isSupporter?: boolean;
-    tickerSettings?: TickerSettings;
-    submitComponentEvent?: (componentEvent: OphanComponentEvent) => void;
-    numArticles?: number;
-    hasOptedOutOfArticleCount?: boolean;
-    email?: string;
-    fetchEmail?: () => Promise<string | null>;
-}
-
-export const bannerSchema = z.object({
-    tracking: trackingSchema,
-    bannerChannel: bannerChannelSchema,
-    content: bannerContentSchema.optional(),
-    mobileContent: bannerContentSchema.optional(),
-    countryCode: z.string().optional(),
-    isSupporter: z.boolean().optional(),
-    tickerSettings: tickerSettingsSchema.optional(),
-    submitComponentEvent: z.any(),
-    numArticles: z.number().optional(),
-    hasOptedOutOfArticleCount: z.boolean().optional(),
-    email: z.string().optional(),
-    fetchEmail: z.any().optional(),
-});
-
-export interface PuzzlesBannerProps extends Partial<BannerProps> {
-    tracking: Tracking;
 }
 
 export interface RawVariantParams {
